refactor: extract modifier helpers
